requirements.txt
streamlit
PyMuPDF
python-docx
scikit-learn
pickle